+++ controllers/clientesController.js
@@ -51,9 +51,14 @@ exports.updateCliente = async (req, res) => {
 exports.deleteCliente = async (req, res) => {
   const { id } = req.params;
   try {
-    const result = await pool.query('DELETE FROM clientes WHERE id = $1 RETURNING *', [id]);
-    if (result.rows.length === 0) return res.status(404).json({ mensagem: 'CLIENTE NÃO ENCONTRADO' });
-    res.json({ mensagem: 'Cliente deletado !' });
+    if (req.usuario.tipo !== 'admin') {
+      return res.status(403).json({ erro: 'Apenas administradores podem remover registros.' });
+    }
+    else {
+      const result = await pool.query('DELETE FROM clientes WHERE id = $1 RETURNING *', [id]);
+      if (result.rows.length === 0) return res.status(404).json({ mensagem: 'CLIENTE NÃO ENCONTRADO' });
+      res.json({ mensagem: 'Cliente deletado !' });
+    }
   } catch (err) {
     res.status(500).json({ erro: err.message });
   }
